Refactor base58 decoding (RIPD-1679):
1) Change the interface to the decoding functions so the caller provides the
buffer that is decoded into.

This allows for an implementation that allocates less memory from the heap, and
potentially makes less memory copies. Since all token types except public keys
have know decoded sizes, it is easy for the caller to provide the buffer. If the
decoded result does not fit exactly in the provided buffer (except for public
keys), the decode fails, as the encoded string cannot represent the expected
token. A flag may be passed to the low-level decoding functions to allow to
decode into a smaller buffer. This is used for public keys, which may decode to
either 32 or 33 bytes.

2) Securely zero memory used in decoding in more situations.

When the low-level decoding routines are decoding a potential secret - like
seeds - then intermediate memory is securely zeroed. When decoding non-secrets -
like account ids - then it does not zero the memory.

3) Use a new algorithm for base 58 decoding.

The algorithm first encodes the base 58 number as a base 58^10 number. All these
coefficients will fit into a 64-bit number. This base 58^10 is then decoded
using boost's multi-precision integer library. This greatly improves decoding
times. My benchmark shows approx 10x improvement on 256-bit decodes, not
including the checksum. An end-to-end test of the old and new decoder shows a
2x-4x improvement, depending on what's being decoded.

4) Decode ripple lib encoded seeds at a low level.

ED25519 encoded with ripple lib used a special encoding. Instead of
<TokenId::FamilySeed><16-bytes seed data>, it uses
<TokenId::None><0xE1><0x4B><16-bytes seed data>. This encoding is handled at the
lowest level. A special flag is passed into the decoder to control if this
decoding should be checked for or not. This simplifies calling client code.
Although it does make testing more complex.