Initial ledger objects and transactions:
Simplified version of the core stable coin transactions and ledger objects. The
audience for this milestone is the rippled developers. In particular, there are
be simplified versions of transactions to:

* Stable coin parameters: Create, delete
* Oracles: Create, delete and update values
* CDPs: Create, delete, deposit, and withdraw
* Stable coins: issue, redeem, and transfer

There is also a new invariant to make sure stable every stable coin in a balance
is tracked in some CDP.

Simplifications:

The following simplifications are made:

* No CDP takeovers
* No buying or selling stable coins
* Oracle updates are unsigned and not a blob (anyone can just update an oracle)
* Single oracle per stable coin
* No transfer fees or time based fees
* No docs
* No RPC support