[fold] Add force_ledger_range_present:
When starting rippled with existing database it do not know it has those
ledgers locally until the server naturally tries to backfill. This makes
is difficult to test some functionality (in particular performance
testing sidechains). With this variable the user is able to force rippled
to consider the ledger range to be present. It should be used for testing
only.